Extract Landmarks to its own file
