+++ Assets/Scripts/Animator/Landmarks.cs
@@ -1,0 +1,51 @@
+// Copyright 2021-2022 The SeedV Lab.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// The names of each BlazePose joints in the same order.
+// https://google.github.io/mediapipe/images/mobile/pose_tracking_full_body_landmarks.png
+public enum Landmarks {
+  Nose = 0,
+  LeftEyeInner = 1,
+  LeftEye = 2,
+  LeftEyeOuter = 3,
+  RightEyeInner = 4,
+  RightEye = 5,
+  RightEyeOuter = 6,
+  LeftEar = 7,
+  RightEar = 8,
+  MouthLeft = 9,
+  MouthRight = 10,
+  LeftShoulder = 11,
+  RightShoulder = 12,
+  LeftElbow = 13,
+  RightElbow = 14,
+  LeftWrist = 15,
+  RightWrist = 16,
+  LeftPinky = 17,
+  RightPinky = 18,
+  LeftIndex = 19,
+  RightIndex = 20,
+  LeftThumb = 21,
+  RightThumb = 22,
+  LeftHip = 23,
+  RightHip = 24,
+  LeftKnee = 25,
+  RightKnee = 26,
+  LeftAnkle = 27,
+  RightAnkle = 28,
+  LeftHeel = 29,
+  RightHeel = 30,
+  LeftFootIndex = 31,
+  RightFootIndex = 32,
+};
