+++ Assets/Scripts/Animator/Landmarks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2bb27c2f4e306443aa1b49d5ebbcfd3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
